refactor：将 chaldea_converter.py 拆分为 chaldea 包，解决 God Module 问题
将 1019 行单文件按职责拆分为 5 个模块：
- servant_types.py: 数据模型 + 常量映射表
- game_data.py: 从者/礼装名称解析（三级 fallback 缓存）
- chaldea_client.py: Chaldea API 通信 + 内容解码 + 输入解析
- bbc_formatter.py: Chaldea → BBC 纯转换逻辑（无 IO 无网络）
- config_checker.py: BBC 配置校验

chaldea_converter.py 保留为向后兼容的薄入口，
bbc_action.py 的调用方式无需改动。
